refactor: improve directory structure
