+++ src/app/private/page.tsx
@@ -1,6 +1,6 @@
 import { redirect } from "next/navigation";
 
-import { createClient } from "@/utils/supabase/server";
+import { createClient } from "@/utils/supabase/auth/server";
 
 export default async function PrivatePage() {
   const supabase = await createClient();
