--- src/app/signup/actions.ts
@@ -1,25 +1,0 @@
-"use server";
-
-import { revalidatePath } from "next/cache";
-import { redirect } from "next/navigation";
-
-import { createClient } from "@/utils/supabase/server";
-
-export async function signup(formData: FormData) {
-  const supabase = await createClient();
-
-  // type-casting here for convenience
-  // in practice, you should validate your inputs
-  const data = {
-    email: formData.get("email") as string,
-    password: formData.get("password") as string,
-  };
-
-  const res = await supabase.auth.signUp(data);
-
-  if (res.error) {
-    redirect("/error");
-  }
-  revalidatePath("/", "layout");
-  redirect("/signin/confirm");
-}
